Added display-product layout
Opted for css grid to have a better control of the UI flow of
display-products. This layout was oriented for mobile version,
so the grid shows two columns of products to let buyer to see
the list of products to buy.. While is responsive design, the
web view is not that important since retailer users
most of the time would use smartphones. The likelihood of using
laptops in a local corner market shop is pretty low. Nerver-
theless, the web view is acceptable,

As a result, the first UI candidate --display-product-- is waiting
to be tested by buyer customer.