Add Retailer Layout
Implemented retailer layout. Retailer and products-by-retailer
attributes were considered to implement this layout. Additionally,
It was necessary to implement  a reusable component. Product-display
was added in shared modules, and it's ready to be used by other
components such as category-componen, type-component and seasonal-
component.

As a result, retailer layout shows the corresponding details of a
retailer store.
.

As a result, just by fedding the componente with two parameters --
typeView and the product as Product() class-- the Product-display
component will show what is needed.