Added Buyer navegation state
Every time the customer buyer navegates in the web app, it's possible
to know where his location is, allowing the app to render the correct
information in the menu bar of the web app.

In short, a dynamic menu bar was implemented. Based on the buyer
location, the app will render the correct feature on the top area of
the web application..