Add Cart Store
In order to have a correct communication between components when
adding new products to shipping cart, a state  management solution
was implemented.

This removes complexity in terms of how to establish comunication
between component from different features (not necessarily child-
parent component comunication).

As a result, updating the shopping Cart store is possible through
different view components in the web app, rendering the current
number of products in the shopping cart by subscribing to
CartStore.