Add deactivate Add component
By clicking the product image, the  add-component modal was disable
from the parent component. This disabeling process is confounf the
customer when closing the modal. Now the add-component modal is
disable from it self by click on the string quantity.

We expect this change to have a high acceptance rate within users.